move styles and types
