+++ packages/confirmation-modal/src/index.ts
@@ -1,1 +1,2 @@
-export { ConfirmationModal as default, Variant } from './ConfirmationModal';
+export { ConfirmationModal as default } from './ConfirmationModal';
+export { Variant } from './ConfirmationModal.types';
